+++ KiwiRichPresence/Configuration/PresenceZoneOverride.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace KiwiRichPresence.Configuration
+{
+    internal class PresenceZoneOverride : PresenceContextOverride
+    {
+        public string Label = string.Empty;
+        public List<uint> TerritoryIds = [];
+    }
+}